requirements.txt
requests
beautifulsoup4
python-telegram-bot==13.15
flask
urllib3<2.0.0
gunicorn